fix(web): extract loadGames from barrel index to fix Turbopack dev resolution
Turbopack cannot resolve named exports from barrel index.ts files in
dev mode. Moving loadGames into a dedicated load-games.ts file fixes
the 'm.loadGames is not a function' TypeError.
