+++ apps/web/src/shared/i18n/messages/index.ts
@@ -95,7 +95,7 @@ export async function loadMessages(locale: Locale): Promise<TranslationBundle> {
   ]);
 
   const [gamesModule, homeData] = await Promise.all([
-    import('./games/index').then(async (m) => ({
+    import('./games/load-games').then(async (m) => ({
       data: await m.loadGames(locale),
     })),
     homeMod.loadHomeMessages(locale),
